requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
groq==0.9.0
PyJWT==2.8.0
bcrypt==4.1.3
supabase==2.4.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1